+++ .gitignore
@@ -1,6 +1,6 @@
 # Archivos que git debería ignorar. Inspirado por el de SciPy.
 # Archivos de PyCharm #
-#########################################
+#######################
 .idea/*
 .settings/
 *.tmp
@@ -10,7 +10,7 @@
 *.py[ocd]
 
 # Archivos de Python #
-################
+######################
 # El directorio de trabajo de setup.py
 build
 # El directorio de sphinx
@@ -20,3 +20,7 @@ doc/_build
 # configuración distutils
 site.cfg
 setup.cfg
+
+# Archivos de VENSIM #
+######################
+.vdf
