+++ tinamit/Ejemplos/Ejemplo básico/Prueba conectado.py
@@ -1,6 +1,6 @@
 import os
 
-from tinamit.NuevoConectado import Conectado
+from tinamit.Conectado import Conectado
 
 
 modelo = Conectado()
@@ -13,4 +13,4 @@ modelo.estab_bf(os.path.join(directorio, 'Prueba bf.py'))
 modelo.conectar(var_mds='Lluvia', var_bf='var1', mds_fuente=False)
 
 print('Simulando...')
-modelo.simular(paso=1, tiempo_final=100, nombre_simul='Corrida_Tinamit')
+modelo.simular(paso=1, tiempo_final=100, nombre_corrida='Corrida_Tinamit')
