+++ tinamit/Ejemplos/Ejemplo básico/Prueba dll.mdl
@@ -75,6 +75,7 @@ $192-192-192,0,Times New Roman|12||0-0-0|0-0-0|0-0-255|-1--1--1|-1--1--1|96,96,1
 ///---\\\
 :L<%^E!@
 9:Current
+23:0
 15:0,0,0,0,0,0
 19:100,0
 27:2,
