+++ tinamit/Interfaz/ControlesGenéricos.py
@@ -9,7 +9,7 @@ except ImportError:
 from . import Arte as Art
 from . import Botones as Bt
 from . import Formatos as Fm
-from .Formatos import gen_formato as gf
+from tinamit.Interfaz.Formatos import gen_formato as gf
 
 
 class ListaItemas(tk.Frame):
